requirements.txt
streamlit>=1.54.0,<2.0.0
python-dotenv>=1.2.1,<2.0.0
pydantic>=2.11.0,<3.0.0

langchain-core>=1.0.0,<2.0.0
langchain-openai>=1.0.1,<2.0.0
langchain-community>=0.4.0,<0.5.0
langchain-experimental>=0.4.0,<0.5.0
langchain-text-splitters>=1.0.0,<2.0.0
langchain-neo4j>=0.6.0,<1.0.0

neo4j>=5.24.0,<7.0.0
faiss-cpu>=1.8.0,<2.0.0
pypdf>=6.0.0,<7.0.0
tiktoken>=0.9.0,<1.0.0
